feat(default-theme): improve performance
